Initial upload of project
Contains functionality for voting with first past the post, approval, and ranked choice voting systems. Data is consumed from a csv file; a sample file in the expected format is provided for reference. The final column being votes of all 0 value is a workaround for a bug causing the last column to not be parsed.

Additional voting systems can be implemented by creating child classes of the Election class and overwriting the DetermineWinner() method. Systems can be added (or removed) from the election by adding (or removing them) from the List<Election> electionTypes, presently found at Program.cs:52.