Improvements focused on errors in ballots
Ballots that contain only 0 votes aren't counted.
Vote entries that aren't numeric values are marked as 0.
Vote entries that are redundant within a ballot are marked as 0. Note that the initial entry of a given value is left as is.
Vote entries that are greater than the amount of candidates are market as 0.
Fix to a minor formatting error in RCV results.